order-service/requirements.txt
fastapi
uvicorn[standard]
pika
tortoise-orm[asyncpg]
aerich